requirements.txt
fastapi
uvicorn[standard]
autopep8
python-dotenv